backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Modular FastAPI backend for Phronesis"
authors = [
    {name = "GenAI Gurus"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"
pytest = "^8.3.5"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-jose = "^3.4.0"
email-validator = "^2.2.0"

